==> code/03.12_Fig_map_soy_ln.Rmd <==
---
title: 'Soy results Map'
author: "Jill Deines"
date: "1/28/2022"
output: 
  html_document:
    toc: true
    toc_float: true
    keep_md: true
---

Goal: make map

Notes:
* gets percentages using the exp since log transformed

**note for published repository:**
data provided in this repository have had latitude/longitude removed to preserve data privacy; in this script, the lat/lon of points are used to average effects per grid cell.

this script retains the code to translate the lat/lon points to the gridded raster, but provides the aggregated rasters only. Code chunks the process the points -> raster are set to EVAl = false. 

**Extra Note:**
When generating the csv of CATE values in the "load model results" chunk for this repo, we were unable to run the predict and average_treatment_effect functions on our causal forests object with grf v 2.0 and greater. Original code was with grf 0.10.2, which is now included with this repo

```{r knitrOpts, echo=FALSE}
library(knitr)
opts_chunk$set(cache=FALSE, fig.path='../../figure/03.12_soyMapFig_latlon/')
```

**R Packages Needed**

```{r packages, message=FALSE, echo=TRUE, eval=TRUE, warning=FALSE}
library(tidyverse)
library(grf)
library(sf)


library(raster)
library(latticeExtra)
library(RColorBrewer)

library(here)

sessionInfo()
```

# Directories

```{r directories}
# scratch folder for model rdata objects
scratchFolder <- '/Users/deinesji/Dropbox/2Stanford/projects/2021_coverCrops/data/causalForests/20220127_scratch_6state_SOY_v04_log_lessVars'

# gis directory
gisFolder <- paste0(here::here(),'/data/gis')

# de-localized CATE output
projDir <- here::here()
dataDir <- paste0(projDir,'/data/cleanedPointSample')
cateFile <- 'causalForestsOutput_soybeans_cates_byState.csv'

# map name
mapName <- 'soy_v04_log_5km'
```

# load model results
and calculate treatment effects as a percent of average yields

## produce CATE file
extract location-free CATEs from causal forest output

```{r loadModel, eval = FALSE}
# dataset
overlap <- readRDS(paste0(scratchFolder, '/overlap.rds'))


Y_hat <- readRDS(paste0(scratchFolder, '/Y_hat.rds'))
cf <- readRDS(paste0(scratchFolder, '/cf_clusters_xc1.rds'))


# process results ------------------------------
# add predicted treatment effects to data 
constOob_predict <- predict(cf)

# add hats to original data frame
overlapTau <- overlap %>%
  bind_cols(data.frame(Y_hat = Y_hat))%>%
  bind_cols(constOob_predict) 

# sampled in analysis
nrow(overlapTau)

# effects in log space
ate_cf_aipw = average_treatment_effect(cf)
ate_cf_aipw = average_treatment_effect(cf)
tauhat_rf_aipw = c(ATE=ate_cf_aipw["estimate"],
                   lower_ci=ate_cf_aipw["estimate"] - 1.96 * ate_cf_aipw["std.err"],
                   upper_ci=ate_cf_aipw["estimate"] + 1.96 * ate_cf_aipw["std.err"])
tauhat_rf_aipw

tauhat_rf_aipw_percent <- (exp(tauhat_rf_aipw) -1) * 100
tauhat_rf_aipw_percent

# display in yield percents
cates <- overlapTau %>%
  mutate(tauPercents =   (exp(predictions) -1) * 100,
         croptype = 'Maize') %>%
  dplyr::select(c(croptype, predictions, tauPercents,latitude, longitude, year, state)) 

write_csv(cates, paste0(dataDir, '/', cateFile))
```

## load and summarize cates

```{r loadCates}
cates <- read_csv(paste0(dataDir, '/', cateFile))


summary(cates$tauPercents)
sum(cates$tauPercents > 0)

# percent with positive effects
sum(cates$tauPercents > 0) / sum(!is.na(cates$tauPercents)) * 100

# get ranges
cates %>%
  summarize(p5 = quantile(tauPercents, probs = c(0.05)),
            p95 = quantile(tauPercents, probs = c(0.95)))

# store output from average_treatment_effect(cf)
tauhat_rf_aipw_percent <- -3.5466
```

# Figure: soy map

## prep: summarize results by grid

writes out raster - not run here

```{r cateMap_makeRasterGrid, eval = FALSE}
# work in aea, epsg:5070
aeaProj <- '+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=23 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +towgs84=1,1,-1,0,0,0,0 +units=m +datum=NAD83 +no_defs'
states_spdf <- spTransform(states_spdf, CRS(aeaProj))

corners <- bbox(states_spdf)
grid <- raster(crs = aeaProj, res = 5000,
                     extent(corners[1,1],     #xmin
                            corners[1,2],      #xmax
                            corners[2,1],       #ymin
                            corners[2,2]))      #ymax
grid[] <- 1:ncell(grid)

gridmask <- mask(grid, states_spdf)
plot(gridmask)
plot(states_spdf, add = TRUE)

# polygonize
gridPoly <- rasterToPolygons(gridmask)

# convert to sf
grid_sf <- st_as_sf(gridPoly) %>% st_transform(4326) %>% rename(cellID = layer)
tau_sf <- st_as_sf(cates, coords = c('longitude','latitude'), crs = 4326)



tau_sf2 <- tau_sf %>%
  filter(croptype == 'Maize') %>%
  dplyr::select(c(year, tauPercents, croptype, predictions)) %>%
  st_join(grid_sf) 

gridTaus0 <- tau_sf2 %>%
  as.data.frame() %>%
  tidyr::drop_na() %>%
  group_by(cellID) %>%
  summarize(meanTauP = mean(tauPercents),
            meanTau = mean(predictions))

gridTaus <- grid_sf %>%
  left_join(as.data.frame(gridTaus0)) %>%
  st_transform(5070)



gridTaus$meanTauP <- replace(gridTaus$meanTauP, gridTaus$meanTauP < -10, -10)
gridTaus$meanTauP <- replace(gridTaus$meanTauP, gridTaus$meanTauP > 10, 10)


taus_spdf <- as(gridTaus, 'Spatial')
tauRas <- rasterize(taus_spdf, gridmask, field = 'meanTauP', background = NA)

writeRaster(tauRas, paste0(gisFolder, '/cateMaps/', mapName, '.tif'))

```


## map vis parameters
for consistent mapping across figures

```{r mapVisParams}
panelBackground <- 'gray30'
stateLines <- 'gray10'
stateFill <- 'gray50'
mapXLim <- c(-90000,1370000)
mapYLim <- c(1470000,2430000)


colorbreaks <- 20

# scale range maize
palMin <- -10
palMax <- 10 

#  color ramps pulled from color brewer (no white)
brownRamp2 <- colorRampPalette(rev(c('#f6e8c3','#dfc27d','#bf812d','#8c510a','#543005')))
greenRamp2 <- colorRampPalette(c('#c7eae5','#80cdc1','#35978f','#01665e','#003c30'))#'

biPal <- c(brownRamp2(colorbreaks/2),greenRamp2(colorbreaks/2))

# state outlines
states <- read_sf(paste0(gisFolder,'/States_continental.shp')) %>%
  st_transform(5070) 
statesStudy <- states %>%  filter(STATE_ABBR %in%             c('SD','MN','IA',"MO",'IL','WI','IN','MI','OH','ND','NE','KS', 'KY'))

states_spdf <- as(statesStudy, 'Spatial')
statesPlot <- as(states, 'Spatial')

# raster map
tauRas <- raster(paste0(gisFolder, '/cateMaps/', mapName, '.tif'))

```


## map

```{r tauMap_soyPercent, fig.width = 6.5, fig.height = 4.5, dpi = 600}
spplot(tauRas, col.regions = biPal, axes = TRUE,
       at=seq(palMin, palMax, (-palMin + palMax)/colorbreaks),
       maxpixels = 150000000,
       colorkey = FALSE,
       xlim = mapXLim, ylim = mapYLim,
       par.settings = list(panel.background = list(col=panelBackground))) +
  latticeExtra::layer(sp.polygons(statesPlot, lwd=.8, col=stateLines)) +
  latticeExtra::layer(sp.polygons(statesPlot, lwd=.8, col=NA, fill = stateFill), under=T)
```




## Histogram scale

```{r cateHist_soy_percent, dev = c('png','pdf'), dpi = 300, fig.width = 2.65, fig.height = 1.75, bg="transparent"}

ggplot(cates,
       aes(x = tauPercents, fill = ..x..)) +
  geom_histogram(breaks = seq(-13, palMax, (-palMin + palMax)/colorbreaks)) +
  xlim(-12.5,3)+
  scale_fill_gradientn(colors = c('#543005','#543005','#543005',biPal)) +
  scale_y_continuous(position = "right") +
  xlab('Yield Impact (%)') +
  ylab('Count') +
  geom_vline(xintercept = 0, col = '#EEECE1', linetype = 'dashed') +
  geom_vline(xintercept = tauhat_rf_aipw_percent, col = 'red', linetype = 'longdash') +
  theme_bw() + theme(panel.grid.major = element_blank(),
                      panel.grid.minor = element_blank(),
                       panel.background = element_rect(fill = "transparent",
                                                     colour = NA),
                     plot.background = element_rect(fill = "transparent",
                                                    colour = NA),
                     panel.border = element_blank(),
                     axis.text = element_text(color = '#EEECE1'),
                     axis.line = element_line(color = '#EEECE1'),
                     axis.title = element_text(color = '#EEECE1'),
                     axis.ticks = element_line(color = '#EEECE1'),
                     legend.position = 'none')


```

### histograms by state

```{r cateHist_byState,  dev = c('png','pdf'), dpi = 300, fig.width = 6.5, fig.height = 1.5}
text_df <- cates %>%
  group_by(state) %>%
  slice(1)


ggplot(cates,
       aes(x = tauPercents)) +
  geom_histogram() +
  xlim(-13,2.5)+
  xlab('Yield Impact (%)') + ylab('Count') +
  geom_vline(xintercept = 0, col = 'black', linetype = 'dashed') +
  geom_vline(xintercept = tauhat_rf_aipw_percent, col = 'red') +
    geom_text(data = text_df, x = -10, y = 1100, size = 3.5, color = 'black',
            aes(label = state)) +
    facet_wrap(~state, nrow = 1) +
  theme_bw()+ theme( strip.background = element_blank(),
                     strip.text = element_blank(),
                     panel.grid.major = element_blank(),
                     panel.grid.minor = element_blank())

  
```


## state inset

```{r stateinset, fig.width = 1.75, fig.height = 1.5, dpi = 300, dev = c('png','pdf')}

plot(states[1], col = 'white', reset = FALSE, main = NA)
plot(statesStudy, col = 'blue', add = TRUE)
```
